Providing feedback
This one was also extremely close. Your buttons are not in the right order as specified in the IO table for the assignment, and your led signal in top does not match to any signals.

I have also added in a bit of code to the test file to make it clearer if you pass or fail.